drm/edid: Drop revision argument for drm_mode_std()
This argument is used to pass in the revision of the EDID, but since the
EDID is passed in as well, we can just as well get the revision from the
EDID structure directly.

While at it, update the kerneldoc comment.

Signed-off-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 464fdeca764bf1dbdf3d21be94ab29054e83a650)

Signed-off-by: Dheeraj Jamwal <dheerajx.s.jamwal@intel.com>